requirements.txt
flask
flask-cors
pymongo
pandas
numpy
xgboost
scikit-learn
python-dotenv
gunicorn